rename files to tsx and add component HamburgerButton
