+++ frontend/src/components/HamburgerButton/HamburgerButton.tsx
@@ -1,0 +1,7 @@
+
+
+export default function HamburgerButton() {
+    return (
+        <div></div>
+    )
+}
