--- frontend/src/components/Header/Header.jsx
@@ -1,14 +1,0 @@
-
-export default function Header() {
-
-
-    return (
-        <header>
-            <HamburgerButton/>
-            <Logo/>
-            <SearchBar/>
-            <ButtonAdd/>
-            <ProfilePicture/>
-        </header>
-    )
-}
