+++ frontend/src/components/ProfilePicture/ProfilePicture.tsx
