+++ frontend/src/components/SearchBar/SearchBar.tsx
